feat: Adds a 'WP is not installed' message for dev environments.
Adds a 'WP is not installed' message for environments labeled as 
'secure,' 'sec,' 'production,' or 'prod.' In these environments, the 
framework will no longer show the normal installation screen. This 
change is essential to prevent inadvertent installations and 
unauthorized takeovers, which can occur for various reasons, such as the 
absence of a connected database or changes to the table prefix. To 
ensure the security of your WordPress installation, make sure to 
explicitly set 'WP_ENVIRONMENT_TYPE' in your .env file to 'staging,' 
'development,' 'dev,' or a similar appropriate value.