feat: add option to Auto-Activate Elementor as Hourly Cron.
This commit introduces a new feature to auto-activate Elementor as an 
hourly cron job. It checks the 'ELEMENTOR_AUTO_ACTIVATION' environment 
variable and, if set to true, attempts to activate Elementor Pro using 
the license provided in 'ELEMENTOR_PRO_LICENSE'. 

If the license is 'valid' and the license data option exists (which 
would assume elementor is already active), the function returns early. 
Otherwise, it attempts to activate Elementor.

This feature is designed to ensure Elementor remains activated in the 
background on an hourly basis.